Improve grabbing video values
Previously if a playlist's first video is private, the entire playlist
would be skipped as the script couldn't pull any video values to know
where to store the files. This would happen sometimes whenever someone
uploads their next video but keeps it private until release day. Now the
script will attempt to grab the value from the 2nd video in a playlist
if the 1st one fails.